REGRESSION (258504@main): SHOULD NEVER BE REACHED in IPC::StreamClientConnection::trySendDestinationIDIfNeeded https://bugs.webkit.org/show_bug.cgi?id=250318 rdar://104027243
Reviewed by NOBODY (OOPS!).

In 258504@main we increased the IPC header size.  It needs to be reflected here too.
This was found by an assertion in GPU process tests.

* Source/WebKit/Platform/IPC/StreamConnectionEncoder.h:

+++ Source/WebKit/Platform/IPC/StreamConnectionEncoder.h
@@ -40,7 +40,7 @@ template<typename, typename> struct ArgumentCoder;
 class StreamConnectionEncoder final {
 public:
     // Stream message needs to be at least size of StreamSetDestinationID message.
-    static constexpr size_t minimumMessageSize = sizeof(MessageName) + sizeof(uint64_t);
+    static constexpr size_t minimumMessageSize = sizeof(MessageName) + sizeof(UInt128);
     static constexpr size_t messageAlignment = alignof(MessageName);
     static constexpr bool isIPCEncoder = true;
